requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
python-telegram-bot[job-queue]==22.0
python-dotenv==1.1.0
httpx==0.28.1
mysql-connector-python==9.3.0
langchain-pinecone==0.2.8
langchain-google-genai==2.1.5
google-genai==1.9.0
pinecone==7.0.2
tenacity==9.1.2